Improve wids function for edge cases and clarify use of extent function
Edge cases are window attribute values that are on the edge of the extent
of a window. So, the lower extent boundary value should lie within the
extent of a given window (inclusive >=) and the upper value should not
(exclusive <). This commit also changes the expectations of tests in that
it acknowledges that the upper boundary value should not be included.